Trv-runtime-into-base (#206)
* feat(manifest): Moving runtime context to base
* feat(compiler): Resolving issue with killing PID of -1
* feat(types): Cleaning up remnant typing isuses

+++ module/rest/src/util/net.ts
@@ -42,7 +42,7 @@ export class RestNetUtil {
       return;
     }
     const [pid] = result.stdout.trim().split(/\n/g);
-    if (pid) {
+    if (pid && +pid > 0) {
       process.kill(+pid);
     }
   }
